cliniscan/requirements.txt
streamlit==1.37.1
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84
matplotlib==3.9.0
tqdm==4.66.4
scikit-learn==1.4.2
kaggle==1.6.14
torch==2.3.1
torchvision==0.18.1
Flask==3.0.3
google-generativeai==0.7.2
langdetect==1.0.9
